# swagger.yaml
openapi: 3.0.1
info:
  title: Blyze
  description: This is the documentation for the backend API supporting Blyze Project.
  contact:
    email: rachitb@gatech.edu
  version: 0.1.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:8080
  description: Localhost
tags:
- name: user
  description: Endpoints related to user management
- name: organizations
  description: Endpoints related to organization management and user linking
- name: projects
  description: Endpoints related to project management and project-organization linking
paths:
  /user:
    put:
      tags:
      - user
      security: []
      summary: Registers a user in the database
      operationId: register
      description: The user must not be already registered in the system. The user must provide at least 8-character password. The user must provide a valid email address. No field can exceed 128 characters.
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/User'
                - $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        200:
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/UserAuthToken'
        400:
          description: Invalid email address supplied
          content: {}
        409:
          description: Email address already in use
          content: {}
        422:
          description: Input rules broken - check description
          content: {}
    post:
      tags:
      - user
      security: []
      summary: Logs in a user
      description: 'Exchanges username and password for a Bearer token. The user must already be registered. The provided combination of email address and password must be correct. No field can exceed 128 characters.'
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/User'
                - $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        200:
          description: User logged in successfully
          content: {}
        401:
          description: The provided combination of email address and password was invalid
          content: {}

  /orgs:
    put:
      tags:
        - organizations
      summary: Creates a new organization with the creator as the owner
      operationId: register_org
      description: The user must be logged in. The organization name must not share any other organization name in the system.
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Organization'
        required: true
      responses:
        200:
          description: Organization Created Successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Organization'
        401:
          description: Authentication token not provided / invalid / expired
        409:
          description: Organization name already exist
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /orgs/{orgId}:
    delete:
      tags:
        - organizations
      summary: Deletes an existing organization
      operationId: delete_org
      description: The user must be already registered in the system. The organization ID supplied must be correct. The user must be the owner of the organization.
      parameters:
        - name: orgId
          in: path
          description: ID of the organization to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Organization Deleted Successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Organization'
        401:
          description: Authentication token not provided / invalid / expired. Also raised when user is not the owner of the organization.
          content: { }
        404:
          description: Invalid organization ID supplied
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /orgs/{orgId}/add/{newUserId}:
    put:
      tags:
        - organizations
      summary: Creates a new organization with the creator as the owner
      operationId: add_user_to_org
      description: The user must be logged in. The logged in user must be the owner of the organization. The new user must not already be in org.
      parameters:
        - name: orgId
          in: path
          description: ID of the organization to which the user must be added
          required: true
          schema:
            type: integer
            format: int64
        - name: newUserId
          in: path
          description: ID of the user that needs to be added
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: User added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Organization'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user is not the owner of the organization.
          content: { }
        404:
          description: Organization not found
          content: { }
        409:
          description: The user is already a member of the given organization
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
    delete:
      tags:
        - organizations
      summary: Removes a user from an organization
      operationId: remove_user_from_org
      description: The user must be logged in. The logged in user must be the owner of the organization. The new user must already be in the organization.
      parameters:
        - name: orgId
          in: path
          description: ID of the organization to which the user must be added
          required: true
          schema:
            type: integer
            format: int64
        - name: newUserId
          in: path
          description: ID of the user that needs to be added
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: User Removed Successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Organization'
        401:
          description: Authentication token not provided / invalid / expired. Also raised when user is not the owner of the organization.
          content: { }
        404:
          description: Invalid organization ID supplied / user not registered in the organization
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /orgs/{orgId}/users:
    get:
      tags:
        - organizations
      summary: Returns a list of users registered in the organization
      operationId: get_users_from_org
      description: The user must be logged in. The user must already be in the organization.
      parameters:
        - name: orgId
          in: path
          description: ID of the organization to which the user must be added
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: List retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UserArray'
        401:
          description: Authentication token not provided / invalid / expired. Also raised when user is not a member of the organization.
          content: { }
        404:
          description: Invalid organization ID supplied
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /projs:
    put:
      tags:
        - projects
      summary: Creates a new organization with the creator as the owner
      operationId: create_proj
      description: The user must be logged in.
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Project'
        required: true
      responses:
        200:
          description: Project Created Successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired
        409:
          description: Project name already exist
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /projs/{projectId}:
    patch:
      tags:
        - projects
      summary: Updates a project
      operationId: update_project
      description: The user must be logged in. The user's organization must be working on the project.
      parameters:
        - name: projectId
          in: path
          description: ID of the project that must be updated
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Project update successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project not found
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
    delete:
      tags:
        - projects
      summary: Deletes a project
      operationId: delete_project
      description: The user must be logged in. The logged in user must be the owner of the organization. The new user must already be in the organization.
      parameters:
        - name: projectId
          in: path
          description: ID of the project that must be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Project deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project not found
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
    get:
      tags:
        - projects
      summary: Returns project info
      operationId: get_project_info
      description: The user must be logged in. The logged in user must be the owner of the organization. The new user must already be in the organization.
      parameters:
        - name: projectId
          in: path
          description: ID of the project that must be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Project info retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project not found
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /projs/{projectId}/orgs/{orgId}:
    patch:
      tags:
        - projects
      summary: Adds an organization to a project
      operationId: add_project_org
      description: The user must be logged in. The user's organization must be working on the project.
      parameters:
        - name: projectId
          in: path
          description: ID of the project to which the organization must be added
          required: true
          schema:
            type: integer
            format: int64
        - name: orgId
          in: path
          description: ID of the organization to be added
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Organization added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project / organization not found
          content: { }
        409:
          description: Organization is already working on the project
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
    delete:
      tags:
        - projects
      summary: Removes an organization from a project
      operationId: remove_project_org
      description: The user must be logged in. The user's organization must be working on the project.
      parameters:
        - name: projectId
          in: path
          description: ID of the project from which the organization must be removed
          required: true
          schema:
            type: integer
            format: int64
        - name: orgId
          in: path
          description: ID of the organization to be removed
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Project deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Project'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project / organization not found
          content: { }
        422:
          description: Input rules broken - check description
          content: { }
  /projs/{projId}/orgs:
    get:
      tags:
        - projects
      summary: Retrieves a list of organizations working on a project
      operationId: get_project_orgs
      description: The user must be logged in. The user's organization must be working on the project.
      parameters:
        - name: projectId
          in: path
          description: ID of the project from which the organization must be removed
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Project organizations successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ProjectArray'
        401:
          description: Authentication token not provided / invalid / expired. Also raised if the logged in user's organization is not working on the project.
          content: { }
        404:
          description: Project
          content: { }
        422:
          description: Input rules broken - check description
          content: { }


components:
  schemas:
    User:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 128
          description: Username, as requested by the user
    UserLogin:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: Password of the user
          format: password
          minLength: 8
          maxLength: 128
    UserAuthToken:
      type: object
      required:
        - auth_token
      properties:
        auth_token:
          type: string
          description: Authentication token of the user
    UserArray:
      type: object
      required:
        - users
      properties:
        users:
          type: array
          description: List of user objects
          items:
            $ref: '#/components/schemas/User'
    Organization:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          description: Internal ID of the organization, as seen in the database
          format: int64
        name:
          type: string
          description: Name of the organization
          minLength: 3
          maxLength: 128
        owner_id:
          type: integer
          description: User ID of the owner of the organization, as seen in the database
          format: int64
    OrganizationArray:
      type: object
      required:
        - orgs
      properties:
        orgs:
          type: array
          description: An array of organization objects
          items:
            $ref: '#/components/schemas/Organization'
    Project:
      type: object
      required:
        - name
        - base_location
      properties:
        id:
          type: integer
          description: Internal ID of the project, as seen in the database
          format: int64
        name:
          type: string
          description: Name of the project
          minLength: 3
          maxLength: 127
        base_location:
          type: string
          description: Main location of the project incident
          minLength: 2
          maxLength: 255
    ProjectArray:
      type: object
      required:
        - projs
      properties:
        projs:
          type: array
          description: An array of project objects
          items:
            $ref: '#/components/schemas/Project'


  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - userAuth: []